requirements.txt
youtube_transcript_api
streamlit
python-dotenv
pathlib